requirements.txt
pyinstaller==6.3.0
Pillow==10.1.0
Pillow==10.1.0
